pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eegstreamer"
version = "0.4.4"
authors = [
    { name="Andrew Mirsky", email="andrew@phenotype.shop" },
    { name="Jack Payne", email="jack@phenotype.shop" },
]
description = "A python package to handle receive, processing and sending eeg data"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta"
]
dependencies = [
    "numpy==1.25.1",
    "scipy==1.11.1",
    "aiofiles==23.1.0",
    "aiohttp==3.8.4",
    "python-osc==1.8.1",
    "pandas==2.0.3",
    "mne==1.4.2",
    "scikit-learn==1.3.0",
    "stringcase==1.2.0",
    "colorama==0.4.6",
    "muselsl==2.2.2"
#    "muselsl @ git+ssh://git@github.com/phenotype-0/muse-lsl.git@main"
]

[project.urls]
"Homepage" = "https://github.com/phenotype-0/example"
"Bug Tracker" = "https://github.com/phenotype-0/issues"

[project.optional-dependencies]
dev = [
    "behave==1.2.6",
    "PyHamcrest==2.0.4"
]

[project.scripts]
oscmonitor = "eegstreamer.__osc_monitor__:main"
eegexample = "eegstreamer.__example__:main"
musemonitor = "eegstreamer.__muse_monitor__:main"

[tool.hatch.metadata]
allow-direct-references = true
